Add files via upload
AutoCAD lisp that gets calculates the force and moment due to wind load on a region with eccentricity from supports.

I used AI - AutoLISP ACE GPT- to code a command that calculates the force and moment on an area due to wind load.  The LISP takes 3 inputs: a REGION, a LINE -representing moment arm- drawn from c.g. of region to the supports, and wind stress value. The lisp writes the values of Force and Moment in MText and positions it in 1/2 and 1/3 of the LINE.



I prompted the GPT to write a code that multiplies the area of a region by the length of a line and then place the result in the midpoint of the line.

After watching the mentioned Autodesk University class -which is a good introduction to Auto LISP-, and prompting the GPT to explain the lisp in details, editing the lisp became easy to ask for wind stress input to calculate the force and added another MText to hold the moment value.